requirements.txt
python-dotenv
langchain
langchain-groq
langchain-core
langchain-community
gradio
